feat: add product deletion endpoints and update Flyway configuration
- Implemented delete functionality in ProductController for single and all products.
- Enhanced ProductService with methods to delete products by ID and delete all products.
- Updated ProductRepository to use Int as the ID type.
- Modified Flyway configuration to disable validation on migration, allowing for checksum mismatches.
- Increased image_url column size in the database migration script.

+++ src/main/kotlin/nl/appetit/api/repository/ProductRepository.kt
@@ -3,6 +3,6 @@ package nl.appetit.api.repository
 import nl.appetit.api.model.Product
 import org.springframework.data.repository.reactive.ReactiveCrudRepository
 
-interface ProductRepository : ReactiveCrudRepository<Product, Long> { }
+interface ProductRepository : ReactiveCrudRepository<Product, Int> { }
